[fix] : 알림 권한 요청 모래사장 -> 로그인 화면으로 변경
