+++ feat/sandbeach/src/main/kotlin/com/team/bottles/feat/sandbeach/SandBeachViewModel.kt
@@ -7,9 +7,6 @@ import com.team.bottles.core.domain.bottle.usecase.GetBottleListUseCase
 import com.team.bottles.core.domain.bottle.usecase.GetPingPongListUseCase
 import com.team.bottles.core.domain.profile.model.UserProfileStatus
 import com.team.bottles.core.domain.profile.usecase.GetUserProfileStatusUseCase
-import com.team.bottles.core.domain.user.model.Notification
-import com.team.bottles.core.domain.user.model.NotificationType
-import com.team.bottles.core.domain.user.usecase.UpdateSettingNotificationUseCase
 import com.team.bottles.exception.BottlesException
 import com.team.bottles.exception.BottlesNetworkException
 import com.team.bottles.feat.sandbeach.mvi.BottleStatus
@@ -25,7 +22,6 @@ class SandBeachViewModel @Inject constructor(
     private val getBottleListUseCase: GetBottleListUseCase,
     private val getPingPongListUseCase: GetPingPongListUseCase,
     private val getRequiredAppVersionUseCase: GetRequiredAppVersionUseCase,
-    private val updateSettingNotificationUseCase: UpdateSettingNotificationUseCase,
     savedStateHandle: SavedStateHandle
 ) : BaseViewModel<SandBeachUiState, SandBeachSideEffect, SandBeachIntent>(savedStateHandle) {
 
@@ -149,19 +145,6 @@ class SandBeachViewModel @Inject constructor(
 
     private fun navigateToPlayStore() {
         postSideEffect(SandBeachSideEffect.NavigateToPlayStore)
-    }
-
-    fun confirmPermission() {
-        launch {
-            NotificationType.entries.forEach { type ->
-                updateSettingNotificationUseCase(
-                    notification = Notification(
-                        notificationType = type,
-                        enabled = true
-                    )
-                )
-            }
-        }
     }
 
 }
